fix(guard): address top-level doctor review
